requirements.txt
faiss-cpu
sentence-transformer
numpy
ddgs
langgraph
langchain
duckduckgo-search
langchain_groq